requirements.txt
fastapi~=0.115.5
pydantic~=2.9.2
pydantic-settings~=2.6.1
starlette~=0.41.2
scikit-learn~=1.5.2
pandas~=2.2.3
numpy~=2.1.3
joblib~=1.4.2
dill~=0.3.9
uvicorn~=0.32.0
aiobotocore~=2.16.1
python-dotenv~=1.0.1